Removes the AbstractPnpInstaller
