+++ packages/plugin-pnp/sources/PnpLinker.ts
@@ -1,13 +1,12 @@
-import {Linker, LinkOptions, MinimalLinkOptions, Manifest, MessageName, DependencyMeta} from '@yarnpkg/core';
-import {FetchResult, Ident, Locator, Package}                                           from '@yarnpkg/core';
-import {miscUtils, structUtils, formatUtils}                                            from '@yarnpkg/core';
-import {CwdFS, FakeFS, PortablePath, npath, ppath, xfs, Filename}                       from '@yarnpkg/fslib';
-import {generateInlinedScript, generateSplitScript, PnpSettings}                        from '@yarnpkg/pnp';
-import {UsageError}                                                                     from 'clipanion';
+import {miscUtils, structUtils, formatUtils, BuildDirective, Descriptor}                                     from '@yarnpkg/core';
+import {FetchResult, Ident, Locator, Package}                                                                from '@yarnpkg/core';
+import {Linker, LinkOptions, MinimalLinkOptions, Manifest, MessageName, DependencyMeta, LinkType, Installer} from '@yarnpkg/core';
+import {CwdFS, FakeFS, PortablePath, npath, ppath, xfs, Filename}                                            from '@yarnpkg/fslib';
+import {generateInlinedScript, generateSplitScript, PackageRegistry, PnpSettings}                            from '@yarnpkg/pnp';
+import {UsageError}                                                                                          from 'clipanion';
 
-import {AbstractPnpInstaller}                                                           from './AbstractPnpInstaller';
-import {getPnpPath}                                                                     from './index';
-import * as pnpUtils                                                                    from './pnpUtils';
+import {getPnpPath, javascriptUtils}                                                                         from './index';
+import * as pnpUtils                                                                                         from './pnpUtils';
 
 const FORCED_UNPLUG_PACKAGES = new Set([
   // Some packages do weird stuff and MUST be unplugged. I don't like them.
@@ -80,17 +79,130 @@ export class PnpLinker implements Linker {
   }
 }
 
-export class PnpInstaller extends AbstractPnpInstaller {
+export class PnpInstaller implements Installer {
   protected mode = `strict`;
 
-  private readonly unpluggedPaths: Set<string> = new Set();
+  private readonly packageRegistry: PackageRegistry = new Map();
+  private readonly blacklistedPaths: Set<PortablePath> = new Set();
 
-  async transformPackage(locator: Locator, manifest: Manifest | null, fetchResult: FetchResult, dependencyMeta: DependencyMeta, {hasBuildScripts}: {hasBuildScripts: boolean}) {
-    if (this.isUnplugged(locator, manifest, fetchResult, dependencyMeta, {hasBuildScripts})) {
-      return this.unplugPackage(locator, fetchResult.packageFs);
-    } else {
-      return fetchResult.packageFs;
+  constructor(private opts: LinkOptions) {
+    this.opts = opts;
+  }
+
+  async installPackage(pkg: Package, fetchResult: FetchResult) {
+    const key1 = structUtils.requirableIdent(pkg);
+    const key2 = pkg.reference;
+
+    const hasVirtualInstances =
+      pkg.peerDependencies.size > 0 &&
+      !structUtils.isVirtualLocator(pkg) &&
+      !this.opts.project.tryWorkspaceByLocator(pkg);
+
+    const manifest = !hasVirtualInstances
+      ? await Manifest.tryFind(fetchResult.prefixPath, {baseFs: fetchResult.packageFs})
+      : null;
+
+    const dependencyMeta = this.opts.project.getDependencyMeta(pkg, pkg.version);
+
+    const buildScripts = manifest !== null && !hasVirtualInstances && !this.opts.project.tryWorkspaceByLocator(pkg)
+      ? javascriptUtils.extractBuildScripts(pkg, fetchResult, manifest, dependencyMeta, {configuration: this.opts.project.configuration, report: this.opts.report})
+      : [];
+
+    const packageFs = !hasVirtualInstances && pkg.linkType !== LinkType.SOFT
+      ? await this.unplugPackageIfNeeded(pkg, manifest, fetchResult, dependencyMeta, {hasBuildScripts: buildScripts.length > 0})
+      : fetchResult.packageFs;
+
+    if (ppath.isAbsolute(fetchResult.prefixPath))
+      throw new Error(`Assertion failed: Expected the prefix path (${fetchResult.prefixPath}) to be relative to the parent`);
+
+    const packageRawLocation = ppath.resolve(packageFs.getRealPath(), fetchResult.prefixPath);
+
+    const packageLocation = normalizeDirectoryPath(this.opts.project.cwd, packageRawLocation);
+    const packageDependencies = new Map<string, string | [string, string] | null>();
+    const packagePeers = new Set<string>();
+
+    // Only virtual packages should have effective peer dependencies, but the
+    // workspaces are a special case because the original packages are kept in
+    // the dependency tree even after being virtualized; so in their case we
+    // just ignore their declared peer dependencies.
+    if (structUtils.isVirtualLocator(pkg)) {
+      for (const descriptor of pkg.peerDependencies.values()) {
+        packageDependencies.set(structUtils.requirableIdent(descriptor), null);
+        packagePeers.add(structUtils.stringifyIdent(descriptor));
+      }
     }
+
+    miscUtils.getMapWithDefault(this.packageRegistry, key1).set(key2, {
+      packageLocation,
+      packageDependencies,
+      packagePeers,
+      linkType: pkg.linkType,
+      discardFromLookup: fetchResult.discardFromLookup || false,
+    });
+
+    if (hasVirtualInstances)
+      this.blacklistedPaths.add(packageLocation);
+
+    return {
+      packageLocation: packageRawLocation,
+      buildDirective: buildScripts.length > 0 ? buildScripts as Array<BuildDirective> : null,
+    };
+  }
+
+  async attachInternalDependencies(locator: Locator, dependencies: Array<[Descriptor, Locator]>) {
+    const packageInformation = this.getPackageInformation(locator);
+
+    for (const [descriptor, locator] of dependencies) {
+      const target = !structUtils.areIdentsEqual(descriptor, locator)
+        ? [structUtils.requirableIdent(locator), locator.reference] as [string, string]
+        : locator.reference;
+
+      packageInformation.packageDependencies.set(structUtils.requirableIdent(descriptor), target);
+    }
+  }
+
+  async attachExternalDependents(locator: Locator, dependentPaths: Array<PortablePath>) {
+    for (const dependentPath of dependentPaths) {
+      const packageInformation = this.getDiskInformation(dependentPath);
+
+      packageInformation.packageDependencies.set(structUtils.requirableIdent(locator), locator.reference);
+    }
+  }
+
+
+  async finalizeInstall() {
+    this.trimBlacklistedPackages();
+
+    this.packageRegistry.set(null, new Map([
+      [null, this.getPackageInformation(this.opts.project.topLevelWorkspace.anchoredLocator)],
+    ]));
+
+    const pnpFallbackMode = this.opts.project.configuration.get(`pnpFallbackMode`);
+
+    const blacklistedLocations = this.blacklistedPaths;
+    const dependencyTreeRoots = this.opts.project.workspaces.map(({anchoredLocator}) => ({name: structUtils.requirableIdent(anchoredLocator), reference: anchoredLocator.reference}));
+    const enableTopLevelFallback = pnpFallbackMode !== `none`;
+    const fallbackExclusionList = [];
+    const fallbackPool = new Map();
+    const ignorePattern = miscUtils.buildIgnorePattern([`.yarn/sdks/**`, ...this.opts.project.configuration.get(`pnpIgnorePatterns`)]);
+    const packageRegistry = this.packageRegistry;
+    const shebang = this.opts.project.configuration.get(`pnpShebang`);
+
+    if (pnpFallbackMode === `dependencies-only`)
+      for (const pkg of this.opts.project.storedPackages.values())
+        if (this.opts.project.tryWorkspaceByLocator(pkg))
+          fallbackExclusionList.push({name: structUtils.requirableIdent(pkg), reference: pkg.reference});
+
+    return await this.finalizeInstallWithPnp({
+      blacklistedLocations,
+      dependencyTreeRoots,
+      enableTopLevelFallback,
+      fallbackExclusionList,
+      fallbackPool,
+      ignorePattern,
+      packageRegistry,
+      shebang,
+    });
   }
 
   async finalizeInstallWithPnp(pnpSettings: PnpSettings) {
@@ -177,17 +289,17 @@ export class PnpInstaller extends AbstractPnpInstaller {
     return nodeModules;
   }
 
-  private async unplugPackage(locator: Locator, packageFs: FakeFS<PortablePath>) {
-    const unplugPath = pnpUtils.getUnpluggedPath(locator, {configuration: this.opts.project.configuration});
-    this.unpluggedPaths.add(unplugPath);
+  private readonly unpluggedPaths: Set<string> = new Set();
 
-    await xfs.mkdirPromise(unplugPath, {recursive: true});
-    await xfs.copyPromise(unplugPath, PortablePath.dot, {baseFs: packageFs, overwrite: false});
-
-    return new CwdFS(unplugPath);
+  private async unplugPackageIfNeeded(locator: Locator, manifest: Manifest | null, fetchResult: FetchResult, dependencyMeta: DependencyMeta, {hasBuildScripts}: {hasBuildScripts: boolean}) {
+    if (this.shouldBeUnplugged(locator, manifest, fetchResult, dependencyMeta, {hasBuildScripts})) {
+      return this.unplugPackage(locator, fetchResult.packageFs);
+    } else {
+      return fetchResult.packageFs;
+    }
   }
 
-  private isUnplugged(ident: Ident, manifest: Manifest | null, fetchResult: FetchResult, dependencyMeta: DependencyMeta, {hasBuildScripts}: {hasBuildScripts: boolean}) {
+  private shouldBeUnplugged(ident: Ident, manifest: Manifest | null, fetchResult: FetchResult, dependencyMeta: DependencyMeta, {hasBuildScripts}: {hasBuildScripts: boolean}) {
     if (typeof dependencyMeta.unplugged !== `undefined`)
       return dependencyMeta.unplugged;
 
@@ -202,4 +314,62 @@ export class PnpInstaller extends AbstractPnpInstaller {
 
     return false;
   }
+
+  private async unplugPackage(locator: Locator, packageFs: FakeFS<PortablePath>) {
+    const unplugPath = pnpUtils.getUnpluggedPath(locator, {configuration: this.opts.project.configuration});
+    this.unpluggedPaths.add(unplugPath);
+
+    await xfs.mkdirPromise(unplugPath, {recursive: true});
+    await xfs.copyPromise(unplugPath, PortablePath.dot, {baseFs: packageFs, overwrite: false});
+
+    return new CwdFS(unplugPath);
+  }
+
+  private getPackageInformation(locator: Locator) {
+    const key1 = structUtils.requirableIdent(locator);
+    const key2 = locator.reference;
+
+    const packageInformationStore = this.packageRegistry.get(key1);
+    if (!packageInformationStore)
+      throw new Error(`Assertion failed: The package information store should have been available (for ${structUtils.prettyIdent(this.opts.project.configuration, locator)})`);
+
+    const packageInformation = packageInformationStore.get(key2);
+    if (!packageInformation)
+      throw new Error(`Assertion failed: The package information should have been available (for ${structUtils.prettyLocator(this.opts.project.configuration, locator)})`);
+
+    return packageInformation;
+  }
+
+  private getDiskInformation(path: PortablePath) {
+    const packageStore = miscUtils.getMapWithDefault(this.packageRegistry, `@@disk`);
+    const normalizedPath = normalizeDirectoryPath(this.opts.project.cwd, path);
+
+    return miscUtils.getFactoryWithDefault(packageStore, normalizedPath, () => ({
+      packageLocation: normalizedPath,
+      packageDependencies: new Map(),
+      packagePeers: new Set<string>(),
+      linkType: LinkType.SOFT,
+      discardFromLookup: false,
+    }));
+  }
+
+  private trimBlacklistedPackages() {
+    for (const packageStore of this.packageRegistry.values()) {
+      for (const [key2, packageInformation] of packageStore) {
+        if (packageInformation.packageLocation && this.blacklistedPaths.has(packageInformation.packageLocation)) {
+          packageStore.delete(key2);
+        }
+      }
+    }
+  }
+}
+
+function normalizeDirectoryPath(root: PortablePath, folder: PortablePath) {
+  let relativeFolder = ppath.relative(root, folder);
+
+  if (!relativeFolder.match(/^\.{0,2}\//))
+    // Don't use ppath.join here, it ignores the `.`
+    relativeFolder = `./${relativeFolder}` as PortablePath;
+
+  return relativeFolder.replace(/\/?$/, `/`)  as PortablePath;
 }
