+++ packages/plugin-pnp/sources/index.ts
@@ -130,7 +130,6 @@ const plugin: Plugin<CoreHooks & StageHooks> = {
   ],
 };
 
-export {AbstractPnpInstaller} from './AbstractPnpInstaller';
 export {PnpInstaller, PnpLinker} from './PnpLinker';
 
 // eslint-disable-next-line arca/no-default-export
